Add files via upload
The board of directors a company that sells motorcycle parts wants to gain a better understanding of wholesale revenue by product line, 
and how this varies month-to-month and across warehouses. Therefore, in this project, i cleaned the sales data, calculated net revenue for each
product line and group the results by month and warehouse.